pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal_manager"
version = "0.1.0"
authors = [
  { name="zhbjsh", email="zhbjsh@outlook.com" },
]
description = "Control and monitor devices with terminal commands"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
keywords = [ 
  "terminal", 
  "command line",
]
dependencies = [
  "python-slugify >= 4.0.1",
]

[project.urls]
"Homepage" = "https://github.com/zhbjsh/terminal-manager"
"Bug Tracker" = "https://github.com/zhbjsh/terminal-manager/issues"